requirements.txt
anthropic>=0.40.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.28.0
reportlab>=4.1.0
sendgrid>=6.11.0
razorpay>=1.4.0
schedule>=1.2.0
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0
sentry-sdk[flask]>=1.40.0
apscheduler>=3.10.0
python-dateutil>=2.9.0
pytz>=2024.1
elevenlabs>=1.0.0
